Revamping search engine filtering
Instead of replacing non-index results with text, the results are now completely hidden. A banner at the top of the page informs the user of the filtering. Users can click buttons to reveal the non-indie results or disable search filtering for that wiki.

Also fixed issue with Ecosia filtering that would conflict with Ecosia's own JavaScript.

Also did some code cleanup and improvements to listeners.