Adjust DokuWiki handing
Renamed platform "doku" to "dokuwiki".
The software is named "DokuWiki", just like "MediaWiki".

The previous search path construction for DokuWiki wikis would fail for
wikis that put their page names in query parameters instead of the path.
For example, https://wiki.diceydungeons.com/doku.php
(note that this wiki is not currently supported by IWB).
The solution is to use "doku.php", which always renders as the main page
 on DokuWikis, regardless of the URL structure, so can safely be used as
the base URL for searches.

However, to parallel MediaWiki wikis, which put "index.php" in the
wiki-specific search_path, I've moved "doku.php" to the definitions of
individual DokuWikis, so the JS now only adds query parameters.