Display all indie wikifarms in content banners
There was a hardcoded check for wiki.gg that would display it in
redirect banners. This has been replaced with a generic notice that
will display any defined host in the redirect banner.

Additionally, the names of hosts in site data are now capitalized
normally (instead of lowercased), so that they can be rendered directly
in text.