Remove reliance on platform to construct URLs (#738)
Refactor code that depends on the destination wiki's platform in order
to support generic wiki platforms (e.g. Wikidot).

Redirection to the Main Page now uses the destination_content_path to
send users directly to the destination wiki's main page, rather than
using a different URL depending on the destination wiki's software.

The only remaining code that depends on the destination wiki's software
is the construction of the search path. This code now has a default case
that doesn't add any software-specific path to a constructed search URL.
This allows any wiki, regardless of the software it runs on, to be
added by putting its full search path in the "destination_search_path"
property of the redirect definition.

As a result of these changes, it is now possible to add redirects to
Wikidot wikis, by just defining the full search path in their redirect
entry's "destination_search_path".

Co-authored-by: SnorlaxMonster <snorlaxmonster@users.noreply.github.com>