Settings: Use 🍃 instead of ༄ for BreezeWiki
Use an actual emoji instead of a Tibetan character to represent
BreezeWiki. Text characters and emoji often render differently (and at
different sizes), so this ensures the BreezeWiki blocks renders more
similarly to the other blocks on the settings page.

Using a semantically-appropriate character is also more accessible for
screenreaders.