Mutation check for removed iwb- classes
On some search engines (like Startpage), the search result DOM is constantly being rewritten and modified, resulting in classes that IWB adds to be removed. This commit adds logic to the mutation observer that checks to see when an iwb- class has been removed, and restores it if it was not a user action (such as clicking to reveal hidden results).