Set background color against immediate parent
The code that was determining the background color of the parent was starting at the parent of the search result. However, the iwb-new-link-container element that this color is being applied to is inserted inside of the search result element, so if the search result itself has a color that is different from its own parent (as is the case on Yandex), this was causing the background color of the iwb-new-link-container to be incorrect.