Repair support for Yandex (#1194)
* Fix selectors on Yandex

The previous selectors were not working on Yandex.
There are a large number of other selectors included, and I do not know the exact circumstances under which they are relevant, so I've left those alone (or made them more broad).

* Set background color against immediate parent

The code that was determining the background color of the parent was starting at the parent of the search result. However, the iwb-new-link-container element that this color is being applied to is inserted inside of the search result element, so if the search result itself has a color that is different from its own parent (as is the case on Yandex), this was causing the background color of the iwb-new-link-container to be incorrect.

---------

Co-authored-by: SnorlaxMonster <snorlaxmonster@users.noreply.github.com>